+++ src/main/java/Glendon.java
@@ -1,16 +1,9 @@
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 import java.util.Scanner;
-import java.util.regex.Pattern;
 
 public class Glendon {
     private static final String name = "Glendon";
 
-    private enum Command {
+    public enum Command {
         BYE("bye"),
         LIST("list"),
         MARK("mark"),
@@ -29,8 +22,6 @@ public class Glendon {
 
     private static final String tasksDataPath = "./data/tasks.txt";
 
-    private static final DateTimeFormatter dateFormat = DateTimeFormatter.ISO_LOCAL_DATE;
-    private static final DateTimeFormatter dateTimeFormat = DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm");
     private final Storage storage;
     private TaskList taskList;
 
@@ -52,11 +43,7 @@ public class Glendon {
 
         while (true) {
             String input = scanner.nextLine();
-            String commandKeyword = input.split(" ")[0];
-            Command command = Arrays.stream(Command.values())
-                    .filter(cmd -> cmd.keyword.equals(commandKeyword))
-                    .findFirst()
-                    .orElse(null);
+            Command command = Parser.parseCommand(input);
 
             if (command == null) {
                 System.out.println("Unknown command");
@@ -72,21 +59,21 @@ public class Glendon {
                 handleList();
                 break;
             case MARK:
-                handleMark(Integer.parseInt(input.split(" ")[1]) - 1);
+                handleMark(Parser.parseIndex(input));
                 saveTasks();
                 break;
             case UNMARK:
-                handleUnmark(Integer.parseInt(input.split(" ")[1]) - 1);
+                handleUnmark(Parser.parseIndex(input));
                 saveTasks();
                 break;
             case DELETE:
-                handleDelete(Integer.parseInt(input.split(" ")[1]) - 1);
+                handleDelete(Parser.parseIndex(input));
                 saveTasks();
                 break;
             case TODO:
             case DEADLINE:
             case EVENT:
-                handleAddTask(input);
+                handleAddTask(Parser.parseTask(input));
                 saveTasks();
                 break;
             default:
@@ -96,7 +83,7 @@ public class Glendon {
     }
 
     public static void main(String[] args) throws GlendonException {
-       new Glendon(tasksDataPath).run();
+        new Glendon(tasksDataPath).run();
     }
 
     private void saveTasks() throws GlendonException {
@@ -126,74 +113,10 @@ public class Glendon {
         System.out.println("Now you have " + this.taskList.getCount() + " tasks in the list.");
     }
 
-    private void handleAddTask(String input) throws GlendonException {
-        try {
-            Task task = taskFromString(input);
-            this.taskList.addTask(task);
-            System.out.println("Got it. I've added this task:");
-            System.out.println(task);
-            System.out.println("Now you have " + this.taskList.getCount() + " tasks in the list.");
-        } catch (GlendonException e) {
-            System.out.println(e.toString());
-        }
-    }
-
-    private Task taskFromString(String str) throws GlendonException {
-        Task task = null;
-        String taskType = str.split(" ")[0];
-        String description;
-        String[] segments;
-        switch (taskType) {
-        case "todo":
-            if (!validateTodoInput(str)) {
-                throw new GlendonException("Invalid todo format");
-            }
-            segments = str.split(" ");
-            description = String.join(" ",
-                    Arrays.copyOfRange(str.split(" "), 1, segments.length));
-            task = new ToDo(description);
-            break;
-        case "deadline":
-            if (!validateDeadlineInput(str)) {
-                throw new GlendonException("Invalid deadline format");
-            }
-            segments = str.split("\\s*(deadline |/by )\\s*");
-            description = segments[1];
-            String dateStr = segments[2];
-            LocalDate date = LocalDate.parse(dateStr, dateFormat);
-            task = new Deadline(description, date);
-            break;
-        case "event":
-            if (!validateEventInput(str)) {
-                throw new GlendonException("Invalid event format");
-            }
-            segments = str.split("\\s*(event |/from |/to )\\s*");
-            description = segments[1];
-            String startStr = segments[2];
-            String endStr = segments[3];
-            LocalDateTime start = LocalDateTime.parse(startStr, dateTimeFormat);
-            LocalDateTime end = LocalDateTime.parse(endStr, dateTimeFormat);
-            task = new Event(description, start, end);
-            break;
-        default:
-            throw new GlendonException("Invalid task format");
-        }
-        return task;
-    }
-
-    private static boolean validateInput(Pattern pattern, String input) {
-        return pattern.matcher(input).matches();
-    }
-
-    private static boolean validateTodoInput(String input) {
-        return validateInput(Pattern.compile("^todo\\s+.+$"), input);
-    }
-
-    private static boolean validateDeadlineInput(String input) {
-        return validateInput(Pattern.compile("^deadline\\s+.+\\s+/by\\s+.+$"), input);
-    }
-
-    private static boolean validateEventInput(String input) {
-        return validateInput(Pattern.compile("^event\\s+.+\\s+/from\\s+.+\\s+/to\\s+.+$"), input);
+    private void handleAddTask(Task task) {
+        this.taskList.addTask(task);
+        System.out.println("Got it. I've added this task:");
+        System.out.println(task);
+        System.out.println("Now you have " + this.taskList.getCount() + " tasks in the list.");
     }
 }
